move bfield options to Run
