+++ Examples/Detectors/MagneticField/CMakeLists.txt
@@ -2,7 +2,6 @@ add_library(
   ActsExamplesMagneticField SHARED
   src/FieldMapRootIo.cpp
   src/FieldMapTextIo.cpp
-  src/MagneticFieldOptions.cpp
   src/ScalableBFieldService.cpp)
 target_include_directories(
   ActsExamplesMagneticField
@@ -12,7 +11,7 @@ target_link_libraries(
   # the ROOT libraries should be private, but if we do that then the linker
   # fails with some missing ROOT symbols.
   PUBLIC ActsCore ActsExamplesFramework ROOT::Core ROOT::Tree
-  PRIVATE Boost::program_options std::filesystem)
+  PRIVATE std::filesystem)
 
 install(
   TARGETS ActsExamplesMagneticField
