+++ Examples/Run/Geant4/DD4hep/DD4hepGeant4Simulation.cpp
@@ -10,8 +10,8 @@
 #include "ActsExamples/DD4hepDetector/DD4hepGeometryService.hpp"
 #include "ActsExamples/DDG4/DDG4DetectorConstruction.hpp"
 #include "ActsExamples/Geant4/Geant4Options.hpp"
-#include "ActsExamples/MagneticField/MagneticFieldOptions.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
+#include "ActsExamples/Options/MagneticFieldOptions.hpp"
 #include "ActsExamples/Simulation/CommonSimulation.hpp"
 
 #include <boost/program_options.hpp>
