+++ Examples/Run/Reconstruction/Common/RecCKFTracks.cpp
@@ -20,11 +20,11 @@
 #include "ActsExamples/Io/Performance/TrackFinderPerformanceWriter.hpp"
 #include "ActsExamples/Io/Root/RootTrajectoryStatesWriter.hpp"
 #include "ActsExamples/Io/Root/RootTrajectorySummaryWriter.hpp"
-#include "ActsExamples/MagneticField/MagneticFieldOptions.hpp"
 #include "ActsExamples/Options/CommonOptions.hpp"
 #include "ActsExamples/Options/CsvOptionsReader.hpp"
 #include "ActsExamples/Options/CsvOptionsWriter.hpp"
 #include "ActsExamples/Options/DigitizationOptions.hpp"
+#include "ActsExamples/Options/MagneticFieldOptions.hpp"
 #include "ActsExamples/Options/SpacePointMakerOptions.hpp"
 #include "ActsExamples/Options/TrackFindingOptions.hpp"
 #include "ActsExamples/Options/TrackFittingOptions.hpp"
